+++ src/iso_date.js
@@ -11,19 +11,19 @@ export default class IsoDate {
   constructor(dateOrYear, month, day) {
     if (dateOrYear && month && day) {
       this.yyyy = dateOrYear.toString()
-      this.mm   = this.#zeroPad(month)
-      this.dd   = this.#zeroPad(day)
+      this.mm   = this.zeroPad(month)
+      this.dd   = this.zeroPad(day)
     } else if (dateOrYear instanceof Date) {
       this.yyyy = dateOrYear.getFullYear().toString()
-      this.mm   = this.#zeroPad(dateOrYear.getMonth() + 1)
-      this.dd   = this.#zeroPad(dateOrYear.getDate())
+      this.mm   = this.zeroPad(dateOrYear.getMonth() + 1)
+      this.dd   = this.zeroPad(dateOrYear.getDate())
     } else if (dateOrYear) {
       [this.yyyy, this.mm, this.dd] = dateOrYear.split('-')
     } else {
       const today = new Date()
       this.yyyy = today.getFullYear().toString()
-      this.mm   = this.#zeroPad(today.getMonth() + 1)
-      this.dd   = this.#zeroPad(today.getDate())
+      this.mm   = this.zeroPad(today.getMonth() + 1)
+      this.dd   = this.zeroPad(today.getDate())
     }
   }
 
@@ -32,21 +32,21 @@ export default class IsoDate {
   }
 
   setDayOfMonth(dayOfMonth) {
-    const date = this.#toDate()
+    const date = this.toDate()
     date.setDate(dayOfMonth)
     return new IsoDate(date)
   }
 
   // @param [Number] first day of the week (Sunday is 0)
   firstDayOfWeek(weekStart) {
-    const date = this.#toDate()
+    const date = this.toDate()
     date.setDate(date.getDate() - (7 + date.getDay() - weekStart) % 7)
     return new IsoDate(date)
   }
 
   // @param [Number] first day of the week (Sunday is 0)
   lastDayOfWeek(weekStart) {
-    const date = this.#toDate()
+    const date = this.toDate()
     date.setDate(date.getDate() + (weekStart + 6 - date.getDay()) % 7)
     return new IsoDate(date)
   }
@@ -102,7 +102,7 @@ export default class IsoDate {
   }
 
   isWeekend() {
-    return [0, 6].includes(this.#toDate().getDay())
+    return [0, 6].includes(this.toDate().getDay())
   }
 
   isToday() {
@@ -110,7 +110,7 @@ export default class IsoDate {
   }
 
   isFirstDayOfWeek(weekStart) {
-    return this.#toDate().getDay() == weekStart
+    return this.toDate().getDay() == weekStart
   }
 
   equals(isoDate) {
@@ -130,13 +130,13 @@ export default class IsoDate {
   increment(unit, count) {
     let date
     if (unit == 'dd') {
-      date = this.#toDate()
+      date = this.toDate()
       date.setDate(date.getDate() + count)
     } else {
       date = unit == 'yyyy'
            ? new Date(+this.yyyy + count, +this.mm - 1)
            : new Date(+this.yyyy, +this.mm - 1 + count)
-      const endOfMonth = IsoDate.#daysInMonth(date.getMonth() + 1, date.getYear())
+      const endOfMonth = IsoDate.daysInMonth(date.getMonth() + 1, date.getYear())
       date.setDate(+this.dd > endOfMonth ? endOfMonth : +this.dd)
     }
     return new IsoDate(date)
@@ -153,7 +153,7 @@ export default class IsoDate {
   static isValidDate(year, month, day) {
     if (year  < 1000 || year  > 9999) return false
     if (month <    1 || month >   12) return false
-    if (day   <    1 || day   > this.#daysInMonth(month, year)) return false
+    if (day   <    1 || day   > this.daysInMonth(month, year)) return false
     return true
   }
 
@@ -162,24 +162,24 @@ export default class IsoDate {
   // @param month [Number] the month (1 is January)
   // @param year [Number] the year (e.g. 2022)
   // @return [Number] the number of days
-  static #daysInMonth(month, year) {
+  static daysInMonth(month, year) {
     if ([1, 3, 5, 7, 8, 10, 12].includes(month)) return 31
     if ([4, 6, 9, 11].includes(month)) return 30
-    return this.#isLeapYear(year) ? 29 : 28
+    return this.isLeapYear(year) ? 29 : 28
   }
 
-  static #isLeapYear(year) {
+  static isLeapYear(year) {
     if ((year % 400) == 0) return true
     if ((year % 100) == 0) return false
     return year % 4 == 0
   }
 
   // Returns a two-digit zero-padded string.
-  #zeroPad(num) {
+  zeroPad(num) {
     return num.toString().padStart(2, '0')
   }
 
-  #toDate() {
+  toDate() {
     // Cannot use `new Date('YYYY-MM-DD')`: it is treated as UTC, not local.
     return new Date(+this.yyyy, +this.mm - 1, +this.dd)
   }