+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 
 ## Unreleased
 
+* Do not use private class methods.
 * Use annotated git tags for releases.
 * Move conditional month logic into model.
 * Refactor some controller internals.